backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
backports-datetime-fromisoformat==2.0.3
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
eventlet==0.40.4
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==3.0.10
flask-marshmallow==1.4.0
Flask-SocketIO==5.6.0
flask-bcrypt-1.0.1 
flask-jwt-extended-4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
pluggy==1.6.0
protobuf==6.33.4
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.47
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
wsproto==1.3.2